pyproject.toml
[tool.ruff]
line-length = 119
target-version = "py39"

[tool.poetry]
name = "lczerolens"
version = "0.1.3"
description = "Interpretability for LeelaChessZero networks."
readme = "README.md"
license = "MIT"
authors = [
    "Yoann Poupart <yoann.poupart@ens-lyon.org>",
]

[[tool.poetry.source]]
name = "PyPI"
priority = "default"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
python-chess = "^1.999"
torch = ">=2"
onnx2torch = "^1.5.13"
tensordict = "^0.3.0"
gradio = {version = "^4.14.0", optional = true}
zennit = "<=0.4.6"
jsonlines = "^4.0.0"
scikit-learn = "^1.4.0"
zennit-crp = "^0.6.0"
einops = "^0.7.0"

[tool.poetry.extras]
demo = ["gradio"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
lczero-bindings = {git = "https://github.com/LeelaChessZero/lc0.git"}
onnxruntime = "^1.16.3"
gdown = "^5.0.1"

[tool.poetry.group.demo]
optional = true

[tool.poetry.group.demo.dependencies]
gradio = "^4.14.0"
lczero-bindings = {git = "https://github.com/LeelaChessZero/lc0.git"}
matplotlib = "^3.8.2"

[tool.poetry.group.doc]
optional = true

[tool.poetry.group.doc.dependencies]
sphinx = "^7.2.6"
pydata-sphinx-theme = "^0.15.2"
sphinx-charts = "^0.2.1"
sphinx-copybutton = "^0.5.2"
plotly = "^5.18.0"
sphinx-design = "^0.5.0"
nbsphinx = "^0.9.3"
pandoc = "^2.3"

[tool.poetry.group.scripts]
optional = true

[tool.poetry.group.scripts.dependencies]
safetensors = "^0.4.2"
pylatex = "^1.4.2"
matplotlib = "^3.8.2"
wandb = "^0.16.3"

[build]
target-dir = "build/dist"
